requirements.txt
langchain
requests
bs4
dotenv
nltk
xml-python
chainlit
tqdm
PyPDF2
textblob
jax==0.4.13
jaxlib==0.4.13
tokenizers==0.19.1
transformers==4.41.0
pinecone
pytz
pymongo
pydantic
bson
bcrypt
uuid
langchain-huggingface
jose
langchain_community
spacy
langchain-pinecone
numpy==1.26.0
google-auth 
google-auth-oauthlib 
google-api-python-client 
dateparser